chore(val-947): refactor modules
