+++ ethereum-validators-set/src/subagents/lido-report/staking-modules/csm-registry.ts
@@ -1,0 +1,7 @@
+import { BaseRegistryModuleContext } from "./base-node-operators-registry";
+
+export class CommunityRegistryModuleContext extends BaseRegistryModuleContext {
+  async fetchOperatorName(operatorId: string, _block: number): Promise<string> {
+    return `CSM Operator #${operatorId}`;
+  }
+}
