+++ ethereum-validators-set/src/subagents/lido-report/staking-modules/curated-registry.ts
@@ -1,0 +1,10 @@
+import { BaseRegistryModuleContext } from "./base-node-operators-registry";
+
+export class CuratedRegistryModuleContext extends BaseRegistryModuleContext {
+  async fetchOperatorName(operatorId: string, block: number): Promise<string> {
+    const { name } = await this.contract.getNodeOperator(operatorId, true, {
+      blockTag: block,
+    });
+    return name;
+  }
+}
